~ renamed _radiogroup to _singleValueGroup to not only support _radio's + Added FONT_System7 and FONT_SystemSymbols8 instead of FONT_Tahoma7 + Added _calendar class! ~ Made some huge speed improvements ~ Made Months Localized ~ Fixed a bug in int2string with the number 0 + Added class _stickybutton ~ Fixed a bug in _bitmap::resize ~ Removed some warnings

+++ include/_library/_fat/fatfile.h
@@ -105,7 +105,7 @@ extern int _FAT_syncToDisc (FILE_STRUCT* file);
 /*
 Method to modify File Attributes
 */
-uint8_t	FAT_getAttr(const char *file);
-int FAT_setAttr(const char *file, uint8_t attr );
+int	FAT_getAttr(const char *file);
+int FAT_setAttr(const char *file, int attr );
 
 #endif // _FATFILE_H
